### 2024-07-28 #7049 @editorialbot
[PRE REVIEW]: PDR: The Planetary Data Reader
**Submitting author:** @Sierra-MC (<a href="http://orcid.org/0000-0001-6065-5461">Sierra V. Brown</a>)
**Repository:** https://github.com/MillionConcepts/pdr
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.1.2
