+++ CHANGELOG.md
@@ -24,7 +24,7 @@
   - Atmospheric Opacity (AO) products
 - MGN RSS solar wind experiment
 - Misc additional products from:
-  - Cassini RPWS and MIMI
+  - Cassini RPWS, MIMI, and ISS
   - Galileo Probe
   - IHW
   - ground-based datasets at SBN
