### 2024-08-07 e1adf52
CI: Create Github Actions workflow that runs unit tests.
Tests are run with CPython 3.9 through 3.12.  Only on Ubuntu for now.
