### 2024-08-07 66bda87
Unit tests: conditionally skip tests that use optional dependencies.
One unit test requires astropy and another one requires PIL.
Skip those tests if the relevant libraries are not installed,
using @pytest.mark.skipif.  Note: this means the tests now have
a hard dependency on pytest, as opposed to other test runners.

Also adjust pdr.browsify to load PIL only in the function that
actually uses it, so that “import pdr.browsify” works when PIL
isn’t installed.  This may mess up the type annotations a little
but they’re already pretty messed up so :shrug:
