### 2024-08-07 9d06ad1
CI: Also test Windows, MacOS, and CPython 3.13.0-rc.1.
CPython 3.13 is scheduled for release in October.
