### 2024-08-07 e93e51b
Unit tests: fix or mask tests that fail only on Windows.
These are unambiguously the correct change given the current state of
PDR proper:

* pdr/tests/utils.py: Don’t write temp files inside pdr/tests/,
  this directory may be read-only in CI.
* pdr/tests/test_fits_load.py: GC the pdr.Data object before
  unlinking the temporary file; on Windows open files cannot be
  deleted.  (TODO: pdr.Data should have a close method.)

This arguably masks a bug in PDR, but that still needs some internal
discussion:

* pdr/tests/objects.py: Strip trailing horizontal white space from
  all strings.
* pdr/te
[…]
