### 2024-08-07 6748fd7
CI: Also test pip-based installation.
There’s no automation keeping the dependencies declared in setup.py
in sync with the dependencies declared in (minimal_)environment.yml,
so let’s at least test in CI that pip-based installs pass the unit
tests.
