### 2024-08-07 aaee4ad
setup.py: Require numpy<2.0.0.
As is already the case in environment.yml.
