### 2024-08-08 7ed1399
Unit tests: Use pytest fixtures for all temporary files and dirs.
This has three benefits: first, no files are ever written to the
source tree, making it possible to run the unit tests from beginning
to end with a read-only pdr source tree (except for __pycache__
directories); second, the tests can be run in parallel (e.g. using
pytest-xdist) without stomping on each others’ temporary files
(this _may_ have been the cause of an intermittent CI test failure)
third, we don’t have to worry about cleaning up any files created
by the tests, because pytest deals with that for us.  (By default,
all t
[…]
