### 2024-08-08 90520ad
CI: Run unit tests with a read-only source tree.
This will catch any future instances of tests and/or pdr core code
being sloppy and writing scratch files relative to either the current
working directory or __file__.  One should instead use pytest’s
per-test scratch directory fixtures.  For examples see commit
7ed13992a3a6559545d44d34870bdfc8b1a97329
(“Unit tests: Use pytest fixtures for all temporary files and dirs”).

Please note that for the CI scripts to work correctly going forward,
the [run] section must be last in `.coveragerc`.  Yes, this is kinda
blech, but there’s no other way to con
[…]
