### 2024-08-11 c18c542
minor style/whitespace changes
* added TYPE_CHECKING-guarded PIL.Image import in browsify
* applied black style to conftest.py
* added some random trailing whitespace back to SILLY_LABEL to annoy the parser
