### 2024-08-11 7bd895d
changelog entries related to new CI (and other light edits)
