### 2024-08-11 5ea2678
basic line ending canonicalization
This is a very simple change that canonicalizes heterodox
line endings in strings or lists/tuples of strings returned to the user
via the `Data.load()` workflow. This has no effect on other types of
Python / PDS objects containing strings, including tables and parsed metadata.
Specifically, it affects only the `ReadText`, `ReadLabel`, and `ReadHeader`
loaders.

It will work reliably only on \r and \r\n endings. Mixed line endings or more
unusual combinations of \r and \n (e.g. \n\r or \r\r\n) may result in extra
line breaks or missing line breaks (which they 
[…]
