### 2024-08-11 ce040c2
fix weird edge case / misleading type hint
