### 2024-08-12 cfe6e0b
Revert mask of Windows-only test failure due to DOS line endings.
As of commit 5ea2678c1582f7f5f9cdaab53cf1814179f38e24 line endings
for text products are canonicalized in PDR core, so the workaround
introduced in e93e51b81f4c8339ca8f91c6184a1387ff642bef should
longer be necessary.
