### 2024-08-19 8714042
Coverage generation tweaks for dashboard project
* Exclude six_new.py from coverage reporting.  It’s a dependency
  of a dependency, not code we care about.
* Generate the coverage-report artifact in Cobertura XML format
  instead of HTML; this will enable the dashboard to show the
  *changes* in coverage from commit to commit.
