### 2024-08-19 9656847
add special case for MSSSO ground data, and update MEX special case hits
