### 2024-08-20 6b02ef0
Convert to pyproject.toml-based packaging.
Declaring metadata of Python packages in setup.py is deprecated
(by the maintainers of distutils and setuptools) in favor of
pyproject.toml.

The _practical_ effect of this change is that ‘pip install -e .’ no
longer issues this complaint:

> DEPRECATION: Legacy editable install of pdr==1.1.2 from
> file:///.../pdr (setup.py develop) is deprecated.  pip 25.0 will
> enforce this behaviour change.  A possible replacement is to add a
> pyproject.toml or enable --use-pep517, and use setuptools >= 64.
> If the resulting installation is not behaving as expe
[…]
