+++ setup.py
@@ -1,40 +1,6 @@
+# All package configuration is now in pyproject.toml.  This file exists
+# solely for backward compatibility, e.g. allowing people to continue
+# to run "python3 setup.py develop" instead of "pip install -e ."
+
 import setuptools
-
-with open("README.md", "r", encoding="utf-8") as fh:
-    long_description = fh.read()
-
-setuptools.setup(
-    name="pdr",
-    version="1.1.2",
-    author="Chase Million",
-    author_email="chase@millionconcepts.com",
-    description="Planetary Data Reader",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    url="https://github.com/MillionConcepts/pdr",
-    packages=setuptools.find_packages(),
-    classifiers=[
-        "Development Status :: 4 - Beta",
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: BSD License",
-        "Operating System :: OS Independent",
-    ],
-    python_requires='>=3.9',
-    install_requires=[
-        "pds4_tools",
-        "multidict",
-        "pandas>=2.0.0",
-        "numpy<2.0.0",
-        "python-Levenshtein",
-        "dustgoggles",
-        "more_itertools"
-    ],
-    extras_require={
-        "notebooks": ["jupyter", "matplotlib"],
-        "browsify": ["matplotlib"],
-        "fits": ["astropy"],
-        "tiff": ["pillow"],
-        "pvl": ["pvl"],
-        "tests": ["pytest"]
-    }
-)
+setuptools.setup()
