### 2024-08-20 60c4a9e
Fix CI failures with pip and python 3.9.
pyproject.toml conversion had the side effect of putting pytest in a
slightly different mode that causes it to trip over the outdated
pds4_tools.extern.six *whether or not* python 3.12 is in use.
(I do not yet know why it trips only when using pip to install
our dependencies, not conda, but it doesn’t really matter.)

While attempting to fix this I discovered *another* bug in pds4_tools:
it declares a dependency on tkinter only if the [viewer] feature is
active, but it loads part of tkinter unconditionally.  (This has also
already been fixed in the pds4
[…]
