+++ pdr/_patches/__init__.py
@@ -1,1 +1,1 @@
-from pdr._patches.patches import patch_pds4_tools_six
+from pdr._patches.patches import patch_pds4_tools
