### 2024-08-20 fe8b2f9
add a special case for MSL REMS RDRs
