+++ .github/workflows/ci-unittest.yml
@@ -7,211 +7,392 @@ on:
       - main
 
 jobs:
+  # Inspired by https://github.com/hynek/build-and-inspect-python-package/
+  # but revised for our needs.  In particular, we have some auxiliary
+  # files, needed for later steps of this process, that don't belong
+  # in the sdist or wheel, that we copy into a (separate) job artifact
+  # so those later steps don't need to check out the source tree.
+  # In the future, we might want to build a conda package as well
+  # as a wheel.
+  build-package:
+    name: Build PDR packages
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out source tree
+        uses: actions/checkout@v4
+        with:
+          path: src
+      # Future: Any sort of validation of the source code itself,
+      # that only needs to run once, could go here.
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        id: py
+        with:
+          python-version: "3.x" # latest stable
+          update-environment: false
+      - name: Prepare build environment
+        run: |
+          '${{ steps.py.outputs.python-path }}' -m venv venv
+          . venv/bin/activate
+          pip install --only-binary :all: --upgrade pip
+          pip install --only-binary :all: \
+              build check-wheel-contents twine wheel
+      - name: Build and verify sdist and wheel
+        working-directory: src
+        run: |
+          . ../venv/bin/activate
+          export SOURCE_DATE_EPOCH=$(git log -1 --pretty=%ct)
+          python -BIm build --outdir=../dist
+          check-wheel-contents ../dist/*.whl
+          python -Im twine check --strict ../dist/*
+      - name: Collect auxiliary files for later stages
+        working-directory: src/.github
+        run: |
+          zip -r ../../dist/build-scripts.zip scripts
+      - name: Upload packages
+        uses: actions/upload-artifact@v4
+        with:
+          name: packages
+          path: dist/pdr*
+
+      - name: Upload build helpers
+        uses: actions/upload-artifact@v4
+        with:
+          name: build-scripts
+          path: dist/build-scripts.zip
+          # this artifact exists only to be input to later jobs,
+          # it does not need to be retained for long
+          retention-days: 1
+
+      - name: Report results
+        run: |
+          ( cd dist
+            printf '\n### Contents of the `packages` artifact\n\n```none\n'
+            shasum -a 256 pdr*
+          ) >> $GITHUB_STEP_SUMMARY
+
   conda-unit:
     name: |
-      ${{ format('conda python: {0} os: {1} env: {2}',
-                 matrix.python-version, matrix.os, matrix.environment-file) }}
+      Test:  ${{matrix.python}}, conda, ${{matrix.os}}, ${{matrix.deps}}
     strategy:
       fail-fast: false
       matrix:
         # Conda-based installations are our recommendation, so we test
         # comprehensively.
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        environment-file: ["environment.yml", "minimal_environment.yml"]
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13.0-rc.1"]
-    runs-on: ${{ matrix.os }}
+        os: [ubuntu, macos, windows]
+        deps: [minimal, full]
+        python: ["3.9", "3.10", "3.11", "3.12"]
+
+    needs: build-package
+    runs-on: "${{ matrix.os }}-latest"
     defaults:
       run:
-        # required for conda environments to work correctly
+        # bash --login is required for conda environments to work correctly;
         # see https://github.com/marketplace/actions/setup-miniconda#important
-        shell: bash -elo pipefail {0}
+        # something in conda's shell scripts is incompatible with -u
+        shell: bash -l -e -o pipefail {0}
     steps:
-      - name: Check out code
-        uses: actions/checkout@v4
[…]
