### 2024-08-22 c789a75
Adjust packaging metadata so sdist and wheel packages are usable.
We were not including all of the files that ought to be included
in PyPI packages.  Corrected.  Regrettably, this cannot be done
using only pyproject.toml, we have to have a MANIFEST.in as well.
(It’s possible that an alternative “build backend,” such as poetry or
hatch, would allow us to avoid this file, but I think that should be
done as a separate work item if we do it at all.)

Because the unit tests ought to be in the sdist but *not* in the
wheel, move them out of the pdr directory.  This has the pleasant
side effect that w
[…]
