### 2024-08-23 feaf116
add support for MEX VMC products
