+++ CHANGELOG.md
@@ -28,6 +28,7 @@ compatibility
   - Atmospheric Opacity (AO) products
 - MGN RSS solar wind experiment
 - MGS RSS solar conjunction data
+- MEX VMC EDRs and RDRs
 - Misc additional products from:
   - Cassini RPWS, MIMI, and ISS
   - Galileo Probe
