### 2024-08-26 55dc2dc
take MEX VMC off the "known unsupported" list
