### 2024-08-26 0137466
add support for Phoenix TEGA led edr and sc rdr
