### 2024-08-30 #7049 @dfm
@Sierra-MC — Thanks for your patience! This summer has been slow for JOSS turnarounds, but we're starting to catch up. I'll take your submission on as editor and start looking for reviewers early next week. Please feel free to share any recommendations for reviewers from your networks or [our database](https://reviewers.joss.theoj.org/) in the meantime.
