### 2024-09-01 092675e
fix bug reading pds4-labeled FITS files
Changes to the loaders.handlers.handle_fits_file() signature were not propagated to the PDS4 FITS-reading
workflow, causing failures when loading some objects. This simply removes the outdated id_as_offset argument
(handle_fits_file() now performs this behavior by default).
