### 2024-09-05 193fe49
add support for more Rosetta CONSERT and MIDAS products
