### 2024-09-06 #7049 @dfm
:wave: @LevN0, @sgillies, @rouault — would any of you be willing to review this submission for The Journal of Open Source Software? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
