### 2024-09-10 7ed53cd
address an edge case in get_bit_start_and_size()
Galileo SSI TELEMETRY_TABLES have a bit column with repeated ITEMS, but mislabel ITEM_BITS as BITS. They are the only example of this so far in the test corpus; future products with repeated bit columns and missing ITEM_BITS can now be handled in check_special_bit_start_case(). Previously line 209 (now 211) threw an error before the special case could be hit.
