### 2024-09-11 96cba2b
Change a BIT_COLUMN edge case fix into a more general special case
Creates a new special case checker for products with BIT_COLUMNs incorrectly defined in an external format file, e.g. Cassini ISS and Galileo SSI telemetry tables. The mistakes in these format files cannot be modified by check_special_block() and errors are hit before check_special_structure() can be called to fix them.
