### 2024-09-13 d8a304b
fix a bug in compute_offsets() and rework the Rosetta MIDAS special case
