### 2024-09-14 #7049 @dfm
👋 @gaelccc — would you be willing to review this submission for The Journal of Open Source Software? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
