### 2024-09-16 fe10a9e
add support for WFF/ATM derived DEM images
