+++ pdr/pds4_tools/pds4_tools/viewer/label_view.py
@@ -1,0 +1,291 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import abc
+
+from .core import Window, SearchableTextWindowMixIn
+from .widgets.tree import TreeView
+
+from ..reader.label_objects import (get_mission_area, get_discipline_area,
+                                    get_spectral_characteristics_for_lid, get_display_settings_for_lid)
+
+from ..extern import six
+from ..extern.six.moves.tkinter import Text, BooleanVar, StringVar
+
+
+@six.add_metaclass(abc.ABCMeta)
+class LabelWindow(SearchableTextWindowMixIn, Window):
+    """ Base class; Window used to display the content of a label """
+
+    def __init__(self, viewer, full_label, structure_label=None, initial_display='full label'):
+
+        # Set initial necessary variables and do other required initialization procedures
+        super(LabelWindow, self).__init__(viewer)
+
+        # Set instance variables
+        self.full_label = full_label
+        self.structure_label = structure_label
+
+        # Stores possibilities of what can be displayed (these are listed in the View menu), 'break' is
+        # not a valid option but simply indicates a break in the menu should be where it occurs
+        self._display_types = ['full label', 'object label', 'break',
+                               'identification area', 'observation area',
+                               'discipline area', 'mission area', 'file area', 'break',
+                               'display settings', 'spectral characteristics']
+
+        # Stores what display type is currently selected
+        display_type = StringVar()
+        self._menu_options['display_type'] = display_type
+        self._add_trace(display_type, 'w', self._update_label, default=initial_display)
+
+        # Stores whether label is being pretty printed or shown as in the file
+        pretty_print = BooleanVar()
+        self._menu_options['pretty_print'] = pretty_print
+        self._add_trace(pretty_print, 'w', self._update_label, default=True)
+
+        # Draw the main window content
+        self._set_heading('Label')
+        self._draw_content()
+
+        # Add notify event for scroll wheel (used to scroll via scroll wheel without focus)
+        self._bind_scroll_event(self._mousewheel_scroll)
+
+        # Add menu options
+        self._add_menus()
+        self._center_and_fit_to_content()
+        self.show_window()
+
+    # Adds menu options used for manipulating the label display
+    def _add_menus(self):
+
+        # Add a File Menu
+        file_menu = self._add_menu('File', in_menu='main')
+        file_menu.add_command(label='Close', command=self.close)
+        file_menu.add_command(label='Close All', command=self._viewer.quit)
+
+        # Add an Edit Menu
+        super(LabelWindow, self)._add_menus()
+
+        # Add View Menu
+        view_menu = self._add_menu('View', in_menu='main')
+
+        for display_type in self._display_types:
+
+            if display_type == 'break':
+                view_menu.add_separator()
+                continue
+
+            view_menu.add_checkbutton(label=display_type.title(), onvalue=display_type,
+                                      offvalue=display_type,
+                                      variable=self._menu_options['display_type'])
+
+            if self._label_for_display_type(display_type) is None:
+                view_menu.entryconfig(view_menu.index('last'), state='disabled')
+
+    # Creates text pad
+    @abc.abstractmethod
+    def _create_text_pad(self, frame):
+        return
+
+    # Sets text shown in text pad for specified label
+    @abc.abstractmethod
+    def _set_label(self, label):
+        return
+
+    # Updates current display to show selected label (based on display type)
+    def _update_label(self, *args):
+
+        display_type = self.menu_option('display_type')
+
+        label = self._label_for_display_type(display_type)
+        self._set_label(label)
+
+    # Obtain the correct label based on the label display type, or None if there is no available label
[…]
