+++ pdr/pds4_tools/pds4_tools/viewer/image_view.py
@@ -1,0 +1,3039 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import math
+import os
+import weakref
+import platform
+import warnings
+from fractions import Fraction
+
+import numpy as np
+import matplotlib as mpl
+from matplotlib.figure import Figure
+
+from . import cache
+from .core import Window, DataViewWindow
+from .mpl import FigureCanvas, MPLCompat
+
+from ..utils.compat import OrderedDict
+from ..utils.helpers import finite_min_max, is_array_like
+from ..utils.exceptions import PDS4StandardsException
+from ..utils.logging import logger_init
+
+from ..extern.zscale import zscale
+from ..extern import six
+from ..extern.six.moves.tkinter import (Frame, Scrollbar, Label, Entry, Scale, Button,
+                                        Radiobutton, StringVar, BooleanVar, DoubleVar, IntVar)
+from ..extern.six.moves.tkinter_tkfiledialog import asksaveasfilename
+
+# Initialize the logger
+logger = logger_init()
+
+#################################
+
+
+class ImageViewWindow(DataViewWindow):
+    """ Window that displays Array_2D and Array_3D PDS4 data structures as an image.
+
+    This window will display image data. After creating it, you should use the load_array() method to load
+    the array that it needs to display.
+    """
+
+    def __init__(self, viewer):
+
+        # Create basic data view window
+        super(ImageViewWindow, self).__init__(viewer)
+
+        # Pack the display frame, which contains the scrollbars and the scrollable canvas
+        self._display_frame.pack(side='left', anchor='nw', expand=1, fill='both')
+
+        # Create the frame which can/will contain the colorbar
+        self._colorbar_frame = Frame(self._static_canvas)
+
+        # Control variable for `freeze_display` and `thaw_display`. Image will not be updated on screen
+        # when counter is greater than 1.
+        self._freeze_display_counter = 0
+
+        # Will be set to an instance of FigureCanvas (containing the main image/slice being displayed)
+        self._figure_canvas = None
+
+        # Will be set to an instance of MPL's AxesImage
+        self._image = None
+
+        # Will be set to an instance of Colorbar
+        self._colorbar = None
+
+        # Will be set to an instance of MPL's toolbar for TK
+        self._toolbar = None
+
+        # Will be set to the masked data array for the structure if it has masked values
+        # (used to minimize operations)
+        self._masked_data = None
+
+        # Will be set to the data for the current slice/image being shown
+        self._slice_data = None
+
+        # Contains sub-widgets of the header
+        self._header_widgets = {'x': None, 'y': None, 'value': None, 'frame': None}
+
+        # Menu option variables. These are TKinter type wrappers around standard Python variables. The
+        # advantage is that you can use trace(func) on them, to automatically call func whenever one of
+        # these variables is changed
+        draw_slice = lambda * args: self._draw_slice()
+        update_colorbar = lambda *args: self._update_colorbar(full_redraw=True)
+        update_invert_axis = lambda *args: self._update_invert_axis(has_default_orientation=False)
+        update_tick_labels = lambda *args: self._update_tick_labels(for_save=False)
+
+        menu_options = [
+            {'name': 'show_colorbar',    'type': BooleanVar(), 'default': True,         'trace': self.toggle_colorbar},
+            {'name': 'colorbar_orient',  'type': StringVar(),  'default': 'horizontal', 'trace': update_colorbar},
+            {'name': 'colormap',         'type': StringVar(),  'default': 'gray',       'trace': self._update_colormap},
+            {'name': 'invert_colormap',  'type': BooleanVar(), 'default': False,        'trace': self._update_colormap},
[…]
