+++ pdr/pds4_tools/pds4_tools/viewer/__init__.py
@@ -1,0 +1,1 @@
+from .core import pds4_viewer
