+++ pdr/pds4_tools/doc/pds4_tools.utils.data_access.rst
@@ -1,20 +1,0 @@
-pds4_tools.utils.data_access module
-===================================
-
-.. currentmodule:: pds4_tools.utils.data_access
-
-Functions
----------
-
-.. autosummary::
-
-    download_file
-    is_supported_url
-    clear_download_cache
-
-Details
--------
-
-.. autofunction:: download_file
-.. autofunction:: is_supported_url
-.. autofunction:: clear_download_cache
