+++ pdr/pds4_tools/doc/pds4_tools.reader.array_objects.rst
@@ -1,34 +1,0 @@
-pds4_tools.reader.array_objects module
-======================================
-
-.. automodule:: pds4_tools.reader.array_objects
-
-Classes
--------
-
-.. autosummary::
-
-    ArrayStructure
-    Meta_ArrayStructure
-    ArraySection
-
-Details
--------
-
-.. autoclass:: ArrayStructure
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
-
-.. autoclass:: Meta_ArrayStructure
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-    .. autoattribute:: id
-
-.. autoclass:: ArraySection
-    :members:
-    :undoc-members:
-    :show-inheritance:
