+++ pdr/pds4_tools/doc/changes.rst
@@ -1,9 +1,0 @@
-:orphan:
-
-.. _changes:
-
-*************
-Release Notes
-*************
-
-.. include:: ../CHANGES.rst
