+++ pdr/pds4_tools/doc/pds4_tools.utils.helpers.rst
@@ -1,23 +1,0 @@
-pds4_tools.utils.helpers module
-===============================
-
-.. currentmodule:: pds4_tools.utils.helpers
-
-Functions
----------
-
-.. autosummary::
-
-    cast_int_float_string
-    is_array_like
-    finite_min_max
-    dict_extract
-    xml_to_dict
-
-Details
--------
-
-.. automodule:: pds4_tools.utils.helpers
-    :members:
-    :undoc-members:
-    :show-inheritance:
