+++ pdr/pds4_tools/doc/pds4_tools.rst
@@ -1,28 +1,0 @@
-:orphan:
-
-pds4_tools package
-==================
-
-Reader
-------
-
-.. toctree::
-    :maxdepth: 2
-
-    pds4_tools.reader
-
-Util
-----
-
-.. toctree::
-    :maxdepth: 2
-
-    pds4_tools.utils
-
-Extern
-------
-
-.. toctree::
-    :maxdepth: 2
-
-    pds4_tools.extern

[…]
