+++ pdr/pdr.py
@@ -310,7 +310,7 @@ class Data:
 
         patch_pds4_tools()
 
-        import pds4_tools as pds4
+        import pdr.pds4_tools as pds4
 
         structure_list = pds4.read(
             self.labelname, lazy_load=True, quiet=True, no_scale=True
@@ -573,7 +573,7 @@ class Data:
         it to our internal FITS-loading workflow.
         """
         structure = self._pds4_structures[object_name]
-        from pds4_tools.reader.label_objects import Label
+        from pdr.pds4_tools.reader.label_objects import Label
 
         if isinstance(structure, Label):
             setattr(self, "label", structure)
