+++ tests/test_parselabel_pds4.py
@@ -1,4 +1,4 @@
-from pds4_tools.reader.label_objects import Label
+from pdr.pds4_tools.reader.label_objects import Label
 from pdr.parselabel.pds4 import reformat_pds4_tools_label
 
 from tests.objects import MINIMAL_PDS4_LABEL
