+++ pdr/pds4_tools/pytest.ini
@@ -5,4 +5,4 @@ norecursedirs =
     .git
     build
     dist
-    pds4_tools/tests/data_writers
+    tests/data_writers
