### 2024-09-17 1ac07d5
remove pds4_tools tests, modify patch to point to internal pds4_tools
