+++ pdr/pds4_tools/tests/data_writers/test_table_writer.py
@@ -1,99 +1,0 @@
-# coding=utf-8
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
-import struct
-import numpy as np
-
-signed_byte =   np.asarray([struct.pack('b', val) for val in [-100, 127, 50]], dtype='object')
-unsigned_byte = np.asarray([struct.pack('B', val) for val in [150, 253, 0]], dtype='object')
-
-signed_msb2 = np.asarray([struct.pack('>h', val) for val in [-32237, 25020, -100]], dtype='object')
-signed_msb4 = np.asarray([struct.pack('>i', val) for val in [2147480000, -1047483647, 143352]], dtype='object')
-signed_msb8 = np.asarray([struct.pack('>q', val) for val in [9003372036854775800, -59706567879, 8379869176]], dtype='object')
-
-unsigned_msb2 = np.asarray([struct.pack('>H', val) for val in [502, 34542, 60535]], dtype='object')
-unsigned_msb4 = np.asarray([struct.pack('>I', val) for val in [50349235, 3994967214, 243414]], dtype='object')
-unsigned_msb8 = np.asarray([struct.pack('>Q', val) for val in [987654073709550582, 25020, 17396744073709550582]], dtype='object')
-
-float_msb =  np.asarray([struct.pack('>f', val) for val in [-1.3862e-43, 1.25e-41, 3.403451e+25]], dtype='object')
-double_msb = np.asarray([struct.pack('>d', val) for val in [1.79e+308, -5.7303e100, -101.432310]], dtype='object')
-
-complex_msb8 = [struct.pack('>f', val) for val in [3.202823*10**38, 1.41*10**5,
-                                                   1.63230, -1.2360*10**10,
-                                                   1.155494*10**-38, -500.23]]
-complex_msb8 = np.asarray([b''.join(complex_msb8[i:i+2]) for i in range(0, 6, 2)], dtype='object')
-
-complex_msb16 = [struct.pack('>d', val) for val in [2.215073858*10**-308, 1.41*10**5,
-                                                    5.072014, -1.2360*10**10,
-                                                    -1.65*10**308, 1.797693*10**308]]
-complex_msb16 = np.asarray([b''.join(complex_msb16[i:i+2]) for i in range(0, 6, 2)], dtype='object')
-
-signed_lsb2 = np.asarray([struct.pack('<h', val) for val in [-32237, 25020, -100]], dtype='object')
-signed_lsb4 = np.asarray([struct.pack('<i', val) for val in [2147480000, -1047483647, 143352]], dtype='object')
-signed_lsb8 = np.asarray([struct.pack('<q', val) for val in [9003372036854775800, -59706567879, 8379869176]], dtype='object')
-
-unsigned_lsb2 = np.asarray([struct.pack('<H', val) for val in [502, 34542, 60535]], dtype='object')
-unsigned_lsb4 = np.asarray([struct.pack('<I', val) for val in [50349235, 3994967214, 243414]], dtype='object')
-unsigned_lsb8 = np.asarray([struct.pack('<Q', val) for val in [987654073709550582, 25020, 17396744073709550582]], dtype='object')
-
-float_lsb =  np.asarray([struct.pack('<f', val) for val in [-1.3862e-43, 1.25e-41, 3.403451e+25]], dtype='object')
-double_lsb = np.asarray([struct.pack('<d', val) for val in [1.79e+308, -5.7303e100, -101.432310]], dtype='object')
-
-complex_lsb8 = [struct.pack('<f', val) for val in [3.202823*10**38, 1.41*10**5,
-                                                   1.63230, -1.2360*10**10,
[…]
