### 2024-09-17 dc31720
remove viewer, cut _patches module, update license
