+++ pdr/pds4_tools/viewer/summary_view.py
@@ -1,1232 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import os
-import io
-import math
-import binascii
-import platform
-import functools
-import traceback
-
-import numpy as np
-
-from . import label_view, table_view, header_view, cache
-from .core import Window, MessageWindow
-from .widgets.scrolled_frame import ScrolledFrame
-from .widgets.tooltip import ToolTip
-from .table_view import array_structure_to_table
-
-from ..reader.core import pds4_read
-from ..reader.data import PDS_array
-from ..reader.read_tables import table_data_size_check
-
-from ..utils.helpers import is_array_like
-from ..utils.logging import logger_init
-
-from ..extern import six
-from ..extern.six.moves.tkinter import (Canvas, Frame, Label, Entry, Button, Radiobutton,
-                                        BooleanVar, StringVar)
-from ..extern.six.moves.tkinter_tkfiledialog import askopenfilename, asksaveasfilename
-
-# Initialize the logger
-logger = logger_init()
-
-#################################
-
-
-class StructureListWindow(Window):
-    """ Window that summarizes the structures showing some of their properties and giving buttons to open them """
-
-    def __init__(self, viewer, quiet=False, lazy_load=False, show_headers=False, withdrawn=False):
-
-        # Set initial necessary variables and do other required initialization procedures
-        super(StructureListWindow, self).__init__(viewer, withdrawn=withdrawn)
-
-        # Set window width to not be resizable
-        self._widget.resizable(width=0, height=1)
-
-        # Initialize structure list window variables
-        self._canvas = None
-        self._scrolled_frame = None
-        self._structure_list = None
-        self._label_open = False
-
-        # Create menu
-        self._add_menus(quiet, lazy_load, show_headers)
-
-        # Add notify event for scroll wheel (used to scroll structure list)
-        self._bind_scroll_event(self._mousewheel_scroll)
-
-    # Opens the label, reads in any structures it contains, calls _draw_summary()
-    def open_label(self, filename=None, from_existing_structures=None):
-
-        if (filename is None) and (from_existing_structures is None):
-            raise TypeError('Cannot open a label without either a filename or existing StructureList.')
-
-        # Reset the summary if a label is already open
-        if self._label_open:
-            self.reset()
-
-        # Open the label and data, or show an error if one occurs
-        try:
-
-            # Lazy-load all structures from file (to obtain meta data)
-            if filename is not None:
-
-                self._structure_list = pds4_read(filename, quiet=self.menu_option('quiet'),
-                                                 lazy_load=True, decode_strings=False)
-
-                cache.write_recently_opened(filename)
-
-            # Load (lazy if previously was lazy) structures from existing list
-            else:
-                self._structure_list = from_existing_structures
-
-            # Set title
-            title = 'Data Structure Summary' if len(self._structure_list) > 0 else 'Label'
-            title += '' if (filename is None) else " for '{0}'".format(filename)
-            self.set_window_title("{0} - {1}".format(self.get_window_title(), title))
-
-        except Exception as e:
-
-            trace = traceback.format_exc()
-            if isinstance(trace, six.binary_type):
-                trace = trace.decode('utf-8')
-
-            log = logger.get_handler('log_handler').get_recording()
-            error = log + '\n' + trace
-            MessageWindow(self._viewer, 'An Error Occurred!', error)
-            logger.exception('An error occurred during label processing.')
-
-        else:
-
-            # Read data from file (or try to access if it exists) for each structure if lazy-load disabled
-            if not self.menu_option('lazy_load'):
[…]
