+++ pdr/pds4_tools/viewer/__init__.py
@@ -1,1 +1,0 @@
-from .core import pds4_viewer
