+++ pdr/pds4_tools/viewer/widgets/tree.py
@@ -1,398 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import textwrap
-import uuid
-
-from ...utils.compat import OrderedDict
-from ...extern import six
-
-
-class TreeView(object):
-    """ A TreeView-like widget for TK. Supports Python 2.6, 2.7 and 3+.
-
-     This widget creates a tree-view, including the ability to minimize and maximize individual leafs.
-     It works by taking in a Text widget, and creating a tree-like view inside of it. You must also input a
-     (potentially nested) `dict` or `OrderedDict` from which the keys and values for the leaves are taken.
-     Each leaf has a key and may have a value. Leaves without any values are considered headers. You may
-     specify the font for headers, keys, values and for the size of the space prior to headers. """
-
-    def __init__(self, text_pad, dictionary,
-                 header_font=None, key_font=None, value_font=None, spacing_font=None):
-
-        self._text_pad = text_pad
-
-        # Add Elements to TreeView
-        self._elements = _make_elements_from_dict(self._text_pad, None, dictionary,
-                                                  header_font=header_font, key_font=key_font,
-                                                  value_font=value_font, spacing_font=spacing_font)
-
-        # Delete first and last line of text pad if these are extra blank lines
-        self._text_pad.config(state='normal')
-
-        for line_index in [['1.0', '2.0'], ['end-1c linestart', 'end']]:
-
-            if self._text_pad.get(*line_index).strip() == '':
-                self._text_pad.delete(*line_index)
-
-        self._text_pad.config(state='disabled')
-
-    @property
-    def text_pad(self):
-        return self._text_pad
-
-    @property
-    def elements(self):
-        return self._elements
-
-    def __len__(self):
-
-        num_elements = 0
-
-        for element in self._elements:
-            num_elements += len(element)
-
-        return num_elements
-
-    def destroy(self):
-        self._text_pad.destroy()
-
-    # Find element in TreeView by element id
-    def find_element_by_id(self, id):
-
-        found_element = None
-
-        for element in self._elements:
-
-            if element.id == id:
-                found_element = element
-
-            else:
-                found_element = element.find_child_by_id(id)
-
-            if found_element is not None:
-                break
-
-        return found_element
-
-
-class Element(object):
-
-    def __init__(self, master, parent, key, value, **kwargs):
-
-        self._id = str(uuid.uuid4())
-
-        self._text_pad = master
-        self._parent = parent
-
-        self._key = key
-        self._value = value
-        self._children = []
-        self._minimized = False
-
-        key, value, children = self._extract_key_value_children()
-
-        # Set element key and value text, if Element has children
-        if children:
-
-            self._set_key_value_text(key='\n', append=True, key_font=kwargs.get('spacing_font'))
-            self._set_key_value_text(key=key, add_newline=True, key_font=kwargs.get('header_font'))
-
-            self._children = _make_elements_from_dict(self._text_pad, self, children, **kwargs)
-
-            self._text_pad.tag_bind(self._id, '<Enter>', lambda e: self._text_pad.config(cursor='arrow'))
-            self._text_pad.tag_bind(self._id, '<Leave>', lambda e: self._text_pad.config(cursor='xterm'))
-            self._text_pad.tag_bind(self._id, '<Button-1>', self.toggle_view)
-
-        # Set element key and value text
-        else:
-
-            self._set_key_value_text(key=key, value=value, add_newline=True,
-                           key_font=kwargs.get('key_font'), value_font=kwargs.get('value_font'))
-
-    @property
-    def id(self):
-        return self._id
-
-    @property
-    def parent(self):
-        return self._parent
-
[…]
