+++ pdr/_patches/__init__.py
@@ -1,1 +1,0 @@
-from pdr._patches.patches import patch_pds4_tools

[…]
