### 2024-09-18 5ef4eec
bump version number to 1.2.0
