+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [X.X.X] - 202X-XX-XX
+## [1.2.0] - 2024-09-18
 ### Added
 
 #### Features
