### 2024-09-18 6bd98cc
update to remove pds4_tools dependencies; edit 1.2.0 release notes
