+++ environment.yml
@@ -4,7 +4,7 @@ channels:
 dependencies:
   - python>=3.9
   - cytoolz
-  - numpy<2.0.0
+  - numpy
   - pandas>=2.0.0
   - git
   - dustgoggles
