### 2024-09-18 0369cfb
bump version number to 1.2.1
