+++ docs/version_history.md
@@ -1,7 +1,22 @@
 # Version History
+## [1.2.1] - 2024-09-18
 ### Added
 
+#### Features
+- Full compatibility with numpy 2.0 for pds4 products 
+(pds3 products already compatible in last release)
+
+### Changed
+- `pds4_tools` is now vendored within `pdr` to apply changes for numpy 2.0 
+compatibility and avoid hot patches
+
+### Removed
+- `pds4_tools` dependency
+- `_patches` module
+
 ## [1.2.0] - 2024-09-18
+### Added
+
 #### Features
 - Compatibility with numpy 2.0 for pds3 products and fits pds4 products
  (non-fits PDS4 product numpy 2.0 compatibility awaits changes in
@@ -11,6 +26,7 @@
 [see issue #57](https://github.com/MillionConcepts/pdr/issues/57)
 - Unit tests are now backed by `pytest` fixtures for improved flexibility and
 compatibility
+- ROW_PREFIX_BYTES are now accounted for when calculating offset positions
 
 #### Dataset Support
 - Galileo NIMS qubes
@@ -79,7 +95,7 @@ compatibility
 - The hotpatch for `pds4-tools` Python 3.12 incompatibility no longer
   has persistent side effects; for instance, `conda` will no longer
   complain about checksum mismatches.
-- bug reading PDS4-labeled FITS files
+- Bug reading PDS4-labeled FITS files
 
 ### Removed
 - `pdr.Data.__iter__` method is being deprecated
