+++ pyproject.toml
@@ -4,7 +4,7 @@
 
 [project]
 name        = "pdr"
-version     = "1.2.0"
+version     = "1.2.1"
 description = "Planetary Data Reader"
 readme      = "README.md"
 license     = { file = "LICENSE.md" }
