### 2024-09-20 8210c70
removed vendored vax code; add rms-vax depedency; NOTE: HAS ASSUMED CONDA PACKAGE IN ENV FILE
