+++ LICENSE.md
@@ -1,6 +1,6 @@
 ###  BSD 3-Clause License
 
-Copyright (c) 2021, Million Concepts
+Copyright (c) 2021-2024, Million Concepts
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -27,24 +27,6 @@ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
 CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-### pdr/vax.py  and pdr/tests/test_vax.py are derived from code in the SETI [pds-tools module](https://github.com/SETI/pds-tools/) and carries this additional license:
-
-Apache License, Version 2.0
-
-Copyright (c) 2023, SETI
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
 
 ### pdr/pds4_tools is derived from code in the Small Bodies Node [pds4_tools package](https://github.com/Small-Bodies-Node/pds4_tools) and carries this additional license:
 
