+++ environment.yml
@@ -12,6 +12,7 @@ dependencies:
   - python-Levenshtein
   - multidict
   - more-itertools
+  - rms-vax
   # optional dependencies (use minimal_environment.yml if you'd prefer not to install)
   - pvl
   - astropy
