+++ LICENSE.md
@@ -1,6 +1,6 @@
 ###  BSD 3-Clause License
 
-Copyright (c) 2021-2024, Million Concepts
+Copyright (c) 2021, Million Concepts
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
