+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - `vax.py` module was removed in favor of using SETI/rms-vax dependency
 
 ### Fixed
-- License copyright extended to 2024; updated docs/LICENSE.md to match LICENSE.md
+- updated docs/LICENSE.md to match LICENSE.md
 
 ### Removed
 
