### 2024-09-20 1a14f16
add a special case for a Cassini ISS calibration ptype
