### 2024-09-21 664e7af
add basic/unstable primary-mode support for desktop image formats
`Data` can now open JPEG, GIF, TIFF, and JP2 files in 'primary' mode.
It unpacks TIFF tags, base- and GPS-range EXIF info, and some basic
image information (mode, dimensions) into the associated Metadata
object. The sole data object associated with the file is always called
'IMAGE'. Not rigorously tested yet.
