### 2024-09-21 ded9853
basic compatibility with geotiff tags
