### 2024-09-21 e20d0e6
don't scale arrays to single value by default
The default dump_browse()/show() behavior applies a clip at the 1st and 99th
percentiles of the array bounds. For some arrays -- for instance an array
that is mostly black with a few small bright regions -- this can result in
an array with only one value. This is undesirable default behavior. This
commit does not apply clipping in this case _unless_ the user has explicitly
specified a clip.
