### 2024-09-21 425cd50
basic support for anigifs
`Data` will now open animated gifs. It loads each frame into a
separate data object, like 'FRAME_0', 'FRAME_1', ..., and loads
the individual frames lazily as usual.
