### 2024-09-21 c93b36a
add support for MPO format images
