### 2024-09-21 de49336
changelog update for primary image support and dump bugfix
