### 2024-09-22 23d73d8
fix issue with baseline JPEGs
