### 2024-09-23 bc2311b
check MIME type of primary desktop images
