### 2024-09-23 1c3d4fe
light refactor of desktop image, add support for still WebP
