+++ pdr/loaders/dispatch.py
@@ -8,7 +8,7 @@ from typing import Optional, TYPE_CHECKING
 from pdr.formats import check_trivial_case
 from pdr.loaders.utility import (
     looks_like_this_kind_of_file,
-    COMPRESSED_IMAGE_EXTENSIONS,
+    DESKTOP_IMAGE_EXTENSIONS,
     FITS_EXTENSIONS,
     IMAGE_EXTENSIONS,
     TABLE_EXTENSIONS,
@@ -42,7 +42,7 @@ def image_lib_dispatch(pointer: str, data: Data) -> Optional[Loader]:
     if looks_like_this_kind_of_file(object_filename, FITS_EXTENSIONS):
         return ReadFits()
     if looks_like_this_kind_of_file(
-        object_filename, COMPRESSED_IMAGE_EXTENSIONS
+        object_filename, DESKTOP_IMAGE_EXTENSIONS
     ):
         return ReadCompressedImage()
     return None
@@ -117,7 +117,7 @@ def file_extension_to_loader(fn: str) -> Loader:
         return ReadText()
     if looks_like_this_kind_of_file(fn, TABLE_EXTENSIONS):
         return ReadTable()
-    if looks_like_this_kind_of_file(fn, COMPRESSED_IMAGE_EXTENSIONS):
+    if looks_like_this_kind_of_file(fn, DESKTOP_IMAGE_EXTENSIONS):
         return ReadCompressedImage()
     return TBD()
 
