### 2024-09-23 0856ac0
add support for animated WebP
