### 2024-09-23 35aa7dc
preserve palette of palette-mode images in metadata
