### 2024-09-24 31d9282
bump the HiRISE RDRs and DTM images up from notional to official support
