### 2024-09-24 #7256 @gaelccc
A few small comments/suggestions while I go through the main example:
1) To run the example notebook `requests` is required. Very easy to install and not a big deal. However given the statement in the manuscript:

> pdr‘s interface is designed to be maximally accessible to the introductory Python user

I would consider adding `requests` to the list of dependencies.

2) For spatial data types (images, raster fields) since the spatial extent/axes is specified in the file labels, it would be a nice addition to be able to dump also the axes, or alternatively have a `.to_xarray()` method which dumps an `xarray.Dataset` with the correct spatial axes.
An example of this would be a geoid error map (`https://pds-geosciences.wustl.edu/mgn/mgn-v-rss-5-gravity-l2-v1/mg_5201/gravity/`).  What I envisage is: besides dumping the array, this addition would allow the user to dump an xarray with fields [geoid_error, longitude, latitude].
