### 2024-09-24 MillionConcepts/pdr#66 @AndrewAnnex
Code of Conduct
As part of openjournals/joss-reviews#7256, as PDR is part of planetarypy I believe the project uses those governance documents but at a glance I don't see a code of conduct markdown file in this repository/mentioned in the Readme.
