### 2024-09-24 MillionConcepts/pdr#67 @AndrewAnnex
make python-Levenshtein optional to allow jupyter-lite compatibility
as part of openjournals/joss-reviews#7256,

suggesting that the python-Levenshtein dependency be made optional to allow compatibility with [jupyterlite](https://jupyterlite.readthedocs.io/en/stable/). I checked the other explicit dependencies listed in the pyproject.toml and they all seemed available. Looking through the repo it also looks like that dependency is only used for one function in the api and it may not be that well used/known.
