### 2024-09-24 MillionConcepts/pdr#68 @Sierra-MC
LRO wea files not parsing correctly
From @gaelccc review

The output I get from reading a wea file (e.g., the ones [here](https://pds-geosciences.wustl.edu/lro/lro-l-rss-1-tracking-v1/lrors_0001/data/wea/)) is wrong. Somehow decimal digits are not assigned to the integer part, but rather to the next columns. E.g., the number 0958.9 is read as 958 and 9 is assigned to the next column. The following column will read 9 N where N is the number supposed to be in that column (e.g., if N=23.4 then 9 N= 9 23). The output type, indeed is str.

[joss-review](https://github.com/openjournals/joss-reviews/issues/7256)
