### 2024-09-24 MillionConcepts/pdr#69 @Sierra-MC
LRO bsp pointers not yet fully implemented
From @gaelccc review

I tried another format mentioned in the example, but officially supported (as per the documentation). Notably bsp files from the LRO repository. In this case the method .read() works, but doesn't seem to produce useful output. All I get is the label of the file and a warning stating that the bsp pointer is still not fully implemented.

[joss-review](https://github.com/openjournals/joss-reviews/issues/7256)
