### 2024-09-24 #7256 @Sierra-MC
> A few comments/suggestions while I go through the main example:
>
> 1. To run the example notebook `requests` is required. Very easy to install and not a big deal. However given the statement in the manuscript:
>
> > pdr‘s interface is designed to be maximally accessible to the introductory Python user
>
> I would consider adding `requests` to the list of dependencies.
>

@gaelccc

We try not to add unnecessary dependencies so we'd prefer to not do that.  `requests` used to back a function we had in `pdr` but was removed as a dependency when we cut those features. I was confused how the binder notebook was running based on your callout that we don't have it as a dependency and did some digging. Apparently `requests` is a dependency of binderhub, so when we removed the dependency we didn't notice an error. We're in the process of considering a move to colab and will be updating the way we run the environments then, in the meantime to fulfill this, I've added a comment in the notebook that explains you'll need to install requests. While we do design pdr to be accessible to introductory Python user, I think this is a reasonable compromise to prevent adding additional dependencies we don't need in the core code.
