### 2024-09-24 9197e8d
remove VICAR reference, fix capitalization typo
