+++ joss_paper/paper.bib
@@ -27,7 +27,7 @@
 
 @misc{PDS4,
       author = {{Jet Propulsion Laboratory}},
-      title = "{Planetary DAta System Standards Reference Version 1.22.0}",
+      title = "{Planetary Data System Standards Reference Version 1.22.0}",
       year = 2024,
       publisher = NASA,
       url = {https://pds.nasa.gov/datastandards/documents/sr/v1/
@@ -90,15 +90,6 @@
       journal = {Github repository},
       title = "{PDS4 Tools}",
       year = 2021
-}
-
-@misc{VICAR,
-      author = {{California Institute of Technology}},
-      url = {https://github.com/NASA-AMMOS/VICAR},
-      publisher = {Github},
-      journal = {Github repository},
-      title = "{VICAR: Video Image Communication and Retrieval}",
-      year = 2023
 }
 
 @misc{GDAL,
