+++ environment.yml
@@ -9,7 +9,7 @@ dependencies:
   - git
   - dustgoggles
   - pip
-  - python-Levenshtein
+  - python-Levenshtein # used only in Metadata.metaget_fuzzy()
   - multidict
   - more-itertools
   # optional dependencies (use minimal_environment.yml if you'd prefer not to install)
