+++ minimal_environment.yml
@@ -8,7 +8,7 @@ dependencies:
   - pandas>=2.0.0
   - git
   - pip
-  - python-Levenshtein # used only in Metadata.metaget_fuzzy()
+  - python-Levenshtein
   - multidict
   - more-itertools
   - dustgoggles
