### 2024-09-24 MillionConcepts/pdr#67 @Sierra-MC
Did a bit of looking into JupyterLite and while it is a cool project it is still unofficial. We also state in our documentation that we only officially support installation into a conda environment which JupyterLite is not compatible with. While there is only the one function in the API that uses it, that function is used heavily is several processing pipelines so we'd rather not leave out the dependency in our recipe.
