### 2024-09-24 MillionConcepts/pdr#68 @cmillion
Hi @gaelccc. We are unable to reproduce the issue as you have described it.

Specifically using `usps_2021279_0648.wea` and `usps_2021279_0648.lbl` and the FMT file pulled [from the archive](https://pds-geosciences.wustl.edu/lro/lro-l-rss-1-tracking-v1/lrors_0001/label/), `pdr` produces the following table:

```
In [17]: d1 = pdr.read('data/usps_2021279_0648.wea')

In [18]: d1['WEAREC_TABLE']
Out[18]:
NAME  HOUR  MINUTE  TEMPERATURE  BAROMETRIC PRESSURE  RELATIVE HUMIDITY  WIND SPEED
0        6      49         15.9                987.5               39.6          29
1        6      59         20.4                985.4               28.0          24
2        7       9         20.4                985.1               28.5          18
3        7      19         20.7                985.1               27.4          25
4        7      29         20.4                985.1               28.3          19
5        7      39         20.5                985.1               28.2          27
6        7      49         20.5                985.1               27.8          17
```

Which appears to agree completely with the values (absent padding zeroes) if I simply open the same wea file up with `vi` e.g.

```
20211006 279 USPS
06:49 015.9 0987.5 039.6 029
06:59 020.4 0985.4 028.0 024
07:09 020.4 0985.1 028.5 018
07:19 020.7 0985.1 027.4 025
07:29 020.4 0985.1 028.3 019
07:39 020.5 0985.1 028.2 027
07:49 020.5 0985.1 027.8 017
```

We confirmed this behavior with several other observations as well.

Can you please provide the specific sequences of commands that you used to generate this strange result and point us to copies of the specific WEA, LBL, and FMT files that you are using?

Crossreferencing: [openjournals/joss-reviews#7256](https://github.com/openjournals/joss-reviews/issues/7256)
