### 2024-09-24 MillionConcepts/pdr#69 @Sierra-MC
@gaelccc could you give an example of the bsp files you're referring to?

If these are the files you're referring to: https://naif.jpl.nasa.gov/pub/naif/pds/data/lro-l-spice-6-v1.0/lrosp_1000/data/spk/

Those are spice files and aren't supported by `pdr`. There is a separate package `spiceypy` that handles spice files (actually authored by our other reviewer Andrew Annex). If these are what you're referring to it's a good catch that we don't directly state we don't support spice kernels and I'll add that to our documentation.
