### 2024-09-24 #7256 @m-stclair
@gaelccc The type of `xarray` interface you describe is a good suggestion, but it is outside the scope of `pdr` for several reasons.

On ARRAY objects and tables:

The spatial metadata of the geoid products you link is very clear on manual inspection, and there are a handful of other products in the PDS with ARRAY objects that follow that _specific_ format (Magellan gravity products and Clementine gravity / topography products). However, Most PDS3 ARRAY objects are not used to describe nice 2D tables, or in general N-d arrays of the kind `xarray` is intended to deal with. They are used to describe repeated 'structs' of various kinds, some but not all of which 'nest' additional arrays, which are sometimes spatial and sometimes not, and their units and intended interpretive layout are often described only in comments, documentation, or adjacent struct fields that have entirely freeform naming conventions. Similarly, many TABLE and SPREADSHEET objects  contain one or more sections that are intended to be interpreted as arrays (which may have multiple columns per array element that are intended to be interpreted as, e.g., x/y/z axes), sometimes supplemented by one or more sections that give units for those arrays, generally alongside columns that give metadata for those arrays. Column names are again entirely freeform, so the relationship between, say, a column that contains reflectance measured at a particular spectrometer bin and a column that contains band center for that bin cannot be straightforwardly determined from automated inspection of the label. Determining in general when these nested objects contain 'axes', what those axes are, what units may or may not be associated with them, and how to break them out and present them to the user would require the application of complex inference heuristics. This is an extremely interesting-sounding project, but it is outside the scope of `pdr`, and might well be as large a work effort as the entirescope of `pdr`.

On rasters:

Some rasters have explicit per-element spatial (or spectral, photometric, etc.) coordinates given in 'backplanes' of some type. This is directly analogous to the ARRAY/TABLE situation above:  the names and locations of these backplanes (often in separate files) are entirely freeform, so automatically determining when 'backplanes' describe multiple physical axes of an array would require application of the same sort of complex inference rules.

Some rasters instead have an associated map projection specification. Some PDS3 map-projected products give one in the label, but others give it only in separate documents / products. Also, many that do contain some kind of projection description in the label do not actually fully specify it, or specify it in a nonstandard way. Many also have variances from the standard that are described only in prose in documentation or catalog files. A few give it only as WCS objects in an associated FITS HDU. This is why even specialty GIS codes like the GDAL PDS3 driver cannot parse a correct projection from many PDS3 products without additional information.

More broadly, computing map projections falls into a category of domain-level processing that is outside the scope of `pdr`. Other major examples include debayering / demosaicking and the application of weights or other calibration factors. The purpose of `pdr` is to reliably put data in a form that can be processed at domain level.
