### 2024-09-24 #7256 @AndrewAnnex
> @AndrewAnnex
>
> > some initial comments on the main paper:
> > lines 44-50: This paragraph mentions two performance claims, and the review guidelines states that performance claims need to be confirmed as per criteria 3 in functionality. Suggesting that a simple benchmark be performed to support.
> > line 45: `pdr` as the first word in the sentence isn't capitalized, now I know the project name is lower case but maybe a filler word could be added to make it clearer where that sentence starts.
> > line 54: same as above line 55: same as above (I stopped here as I noticed some instances at earlier lines)
> > line 61-64: VICAR and ISIS aren't python packages so not sure if it is really relevant to the subject of Python libraries for accessing PDS data. They certainly enable access to PDS data (to varying degrees) but they fundamentally have different goals/aims. Comparisons to pds4_tools, astropy, etc are more aligned with this. Suggesting either not mentioning these other projects or qualifying them by stating they don't attempt to address the issue of access of PDS data with Python explicitly.
> > line 80: Typo in capitalization of title
>
> How should we include the benchmarks? We have some that we've run that would be appropriate to support those claims, but I'm just unsure what's the usual format/method that they're shown here. Is that something that just is presented during this review, that is included as a reference that we put on zenodo, that is stated in the text, or perhaps some other method? @dfm do you have clarification on this?
>
> For starting sentences with proper lowercase it appears to be based on style guides. It's something Million Concepts does quite often, letting the period mark where a new sentence starts. If JOSS has a particular style-guide convention banning this I'll reword, but if not, I'd prefer to leave it as is as to not introduce awkwardly worded sentences or disrupt the flow solely for capitalization purposes. @dfm I don't see a style guide on the website, does JOSS have one?
>
> The VICAR/ISIS mention (and reference) has been removed.
>
> The typo in title capitalization has been fixed. Good catch.

I don't think there is an accepted standard for the benchmarks you need to worry about, as a suggestion something like the time required to parse a single label for example (with the units) would be sufficient.

As I read it, the reviewer's checklist says the justification should be made within the manuscript text and it should referencing specific performance metrics. Its is up to authors to devise the text and the metrics to present to justify performance claims (and the reviewer to judge if it's sufficient). Otherwise removing mentions of performance/speed solves the problem perhaps in a simpler manner.
