### 2024-09-24 MillionConcepts/pdr#67 @AndrewAnnex
@Sierra-MC what do you mean by "unofficial" here? As pdr doesn't implement any c level code, by making the dependency optional you instantly gain compatibility with pyodide with no other work needed beyond the less than 5 lines of code changes necessary (it may literally be 2). Also, you do support setuptools/pip installs via the pyproject.toml and you distribute PDR on PyPI so I don't think it is true to say conda is the only officially supported way to install PDR. I don't think this issue is a blocker for the JOSS review, it is just a quick suggestion that I suspect will make pdr even more accessible
