### 2024-09-24 MillionConcepts/pdr#67 @cmillion
We support pip install as a convenience. We only officially support install via conda-forge in a `pdr` conda environment.
