### 2024-09-24 MillionConcepts/pdr#67 @Sierra-MC
@AndrewAnnex I wasn't familiar with JupyterLite so I went to [their webpage](https://jupyterlite.readthedocs.io/en/stable/). The section on status states "Although JupyterLite is currently being developed by core Jupyter developers, the project is still unofficial."

So by "unofficial" I'm referring to that statement specifically, not like, any particular judgement on it. Like I said this is the first I'm hearing of the project so far be it from me to judge their status. I looked at their issues and they do have an open issues to be able to check if packages are compatible (https://github.com/jupyterlite/jupyterlite/issues/967) and to better document the limitations (https://github.com/jupyterlite/jupyterlite/issues/199). Without familiarity I wasn't entirely positive what made Levenshtein incompatible (though you've clarified here it is the C code which is what I suspected from the bit of poking I did do). The team discussed the possibility of changing the dependency to optional in the project.toml, however this would make it auto-install for conda users and not for pip users. Which is a little confusing behavior and potentially difficult to diagnose.

I'm not really sure how big of a market the pyodide community represents so it's difficult to assess the degree of accessibility that this opens and weigh that against the possible confusion the different installation behavior would cause.
