### 2024-09-24 MillionConcepts/pdr#68 @gaelccc
Thank you for the prompt reply.
Indeed trying on this file works well for me too.
I incurred in this issue while trying to read an older file: `https://pds-geosciences.wustl.edu/lro/lro-l-rss-1-tracking-v1/lrors_0001/data/wea/lro_es_107/2021279/ku2s_2013070_0916.wea` and associated `.lbl`.
I am using the same .FMT (`lro_wea_rec_ws3.fmt`) file for both cases.
This is the output I get for this minimal example:

```
fname = 'ku2s_2013070_0916.lbl'
data=pdr.read(fname)
data['WEAREC_TABLE']

OUT:
NAME  HOUR  MINUTE  TEMPERATURE  BAROMETRIC PRESSURE RELATIVE HUMIDITY  \
0        9      16         -8.0                  958             9 057
1        9      21         -8.0                  958             9 057
2        9      26         -8.0                  958             9 057
3        9      31         -8.0                  958             8 056
4        9      36         -8.0                  958             8 055
5        9      41         -8.0                  958             8 056
6        9      46         -8.0                  958             8 056

NAME WIND SPEED
0           1 0
1           3 0
2           0 0
3           2 0
4           7 0
5           4 0
6           NaN

```

This is what the `.wea` file looks like:

```
20130311 070 KU2S
09:16  -08.4  0958.9 057.1 011
09:21  -08.3  0958.9 057.3 008
09:26  -08.2  0958.9 057.0 007
09:31  -08.1  0958.8 056.2 011
09:36  -08.4  0958.8 055.7 011
09:41  -08.4  0958.8 056.4 008
09:46  -08.4  0958.8 056.4 010
```
