### 2024-09-24 MillionConcepts/pdr#69 @gaelccc
Apologies! While copy/pasting my comments I omitted a fundamental "not":

> I tried another format **not** mentioned in the example, but officially supported (as per the documentation)

Yes, I am referring to BSP/SPK files. I wasn't expecting `pdr` to support them, but since the [supported datasets](https://github.com/MillionConcepts/pdr/blob/main/docs/supported_datasets.md) document states in few instances _trajectory data_ (e.g., Galileo), I just naively assumed it referred to `BSP`.

Now after a deeper look I see that the _trajectory data_ appears only on specific missions, in which case I assume the archivers decided to use a different format and not BSP, that is the one you are referring to.

I think this issue can easily be closed by just specifying somewhere something like: `trajectory data (no BSP/SPK)`
