### 2024-09-24 #7256 @gaelccc
@m-stclair Thank you for the thorough explanation. I understand now the complexity and agree that the correct interpretation of domain/axes can be left to the data user.

@Sierra-MC I understand the reason for minimizing dependencies. I find the solution of adding a comment in the notebook sufficient and appropriate.
