### 2024-09-24 MillionConcepts/pdr#70 @gaelccc
Community guidelines" need some text
Cross-referencing [openjournals/joss-reviews#7256](https://github.com/openjournals/joss-reviews/issues/7256)

Regarding the **Community guidelines**, namely

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I really like the google form for submitting a request for additional data types.

Points 1, 2 and 3, however, are not explicitly addressed. I suggest to add a couple of lines in the `README` to address these points. What is the contribution philosophy? Is the preferred method of interaction issues/PRs on github or do you prefer users to directly contact one of the main developers?
