### 2024-09-25 3c70d1f
delay PIL import correctly, move changelog to actual changelog file
