+++ CHANGELOG.md
@@ -2,17 +2,39 @@
 ### Added
 
 #### Features
-
+- Support for 'desktop' image formats in primary mode: JPEG, TIFF, GIF 
+  (still and animated), PNG, MPO/MPF, WebP (still and animated), JPEG2000/JP2, 
+  and BMP. Behavior is as follows:
+  - `pdr` will attempt to parse all EXIF, TIFF tag, and MPO metadata, including
+  embedded XML packets. Parsed metadata is available in `Data.metadata` and 
+  accessible with `Data.metaget()` and related methods.
+  - Some basic image information is also available in `Data.metadata` (format, 
+  color mode, dimensions, MIME type, etc.) 
+  - Most files will have a single data object named 'IMAGE'.
+  - Animated GIF and WebP images will have one data object per frame, named 
+    'FRAME_0', 'FRAME_1', etc.
+  - MPO/MPF files (this includes many files with .jpg and .jpeg extensions)
+  will have one data object per image. The MPO primary image is named 'IMAGE'. 
+  Subsequent images are assigned names based on their MPType and index within
+  the file.
+  - PDR automatically converts palette-mode images (like many GIFs) to 3D RGB
+  arrays on load. In these cases, mappings between RGB values and original 
+  palette indices are available with `data.metaget('palette')`.
+  - GeoTiffs work like other TIFF files. `pdr` naively parses the GeoTiff 
+  metadata, but does not construct a CRS or otherwise derive projection 
+  information.
 
 #### Dataset Support
 - additional Cassini ISS calibration images
 - MRO HiRISE RDRs and additional DTM images (were notionally supported, now have official support)
+  `image_clip` argument will override this, if for some reason you want to.
 
 ### Changed
-
+- `Data.dump_browse()` and `.show()` no longer apply default contrast enhancement 
+  if it would result in an image with only one value. Explicitly passing the
 
 ### Fixed
-
+- `Data.dump_browse()` now works correctly with FITS files opened in primary mode.
 
 ### Removed
 
