### 2024-09-25 e7a94c9
remember that python 3.9 does not have match/case statements
