### 2024-09-25 MillionConcepts/pdr#68 @m-stclair
Thanks @gaelccc. We looked into this and discovered that the ku2s* files in that directory have bad metadata. Like the usps* files Sierra linked, these ku2s* files specify the LRO_WEA_REC_WS3.FMT file for table structure, and state in their primary labels that they have 29 bytes per row. However, the ku2s* files actually have 32 bytes per row, and their column boundaries do not match the column descriptions in the referenced FMT file.

We have not checked to see if this is consistent across all ku2s* files in the LRORS corpus. If it is, we could write a special case for this. However, when we discover data quality issues related to an ongoing mission, we prefer to first reach out to the data providers to let them know about the problem and see if they have the capability to fix it. Do you happen to know the appropriate person to contact about these files?
