### 2024-09-25 fb49080
add information on spice files
