+++ docs/supported_datasets.md
@@ -44,7 +44,7 @@ or .spk from NAIF), in part because "opening" a spice file in the way `pdr`
 reads planetary data isn't very useful. If you're looking to work with these 
 files we recommend [spiceypy](https://github.com/AndrewAnnex/SpiceyPy) or 
 [lhorizon](https://github.com/MillionConcepts/lhorizon) (which works with spice
-kernels in combination with JPL Horizons.
+kernels in combination with JPL Horizons).
 
 ## Officially Supported Datasets:
 #### Apollo
