### 2024-09-25 b02b0a0
add community guidelines info
