### 2024-09-25 MillionConcepts/pdr#66 @Sierra-MC
Code of conduct added in [this commit](https://github.com/MillionConcepts/pdr/commit/b02b0a038f167d50e1b29073896305439f1ec483). I used an updated version of the code used by planetarypy
