### 2024-09-26 MillionConcepts/pdr#68 @Sierra-MC
After looking into this further I've confirmed the number of bytes per row is not consistent throughout all the ku2s* files. We can implement a special case that simply ignores the format files and parses as delimiter separated tables; preference is still to bring this to the attention of the data providers and have them implement a fix.
