### 2024-09-26 33e0c78
make Levenshtein an optional dependency
