+++ pyproject.toml
@@ -31,7 +31,6 @@ dependencies = [
     "multidict",
     "numpy",
     "pandas>=2.0.0",
-    "python-Levenshtein",
 ]
 
 [project.optional-dependencies]
@@ -41,6 +40,7 @@ notebooks  = ["jupyter", "matplotlib"]
 pvl        = ["pvl"]
 tests      = ["pytest"]
 tiff       = ["pillow"]
+fuzzy      = ["Levenshtein"]
 
 [project.urls]
 Repository = "https://github.com/MillionConcepts/pdr"
