### 2024-09-26 MillionConcepts/pdr#67 @Sierra-MC
This has continued to eat at me the past couple of days and after further discussion we have decided to make the change you suggested @AndrewAnnex . JupyterLite may end up being what we use to replace binder, but even if not you've made good points that we can allow further compatibility with pyodide. This was changed in [this commit](https://github.com/MillionConcepts/pdr/commit/33e0c789c050e2dbcf57f646fed15c55ee1debe0) on `develop` so it will be compatible in our next release. Thank you for the suggestion and for bringing this package to our attention.
