### 2024-09-26 2445b52
update tests to for optional Levenshtein dependency
