### 2024-09-26 a4a6e36
install fuzzy optinal dependency during ci run
