### 2024-09-26 0a95bc6
and remove Lev from conda minimal env :eyes:
