### 2024-09-26 MillionConcepts/pdr#68 @gaelccc
Indeed, I think it's better to inform the data provider. I sent a message to one of them, let's see.
Thanks for digging into this.
I think that it would be good, generally speaking, to have some control mechanism to at least verify that the parsed numbers make sense (i.e., the `int()` or `float()` operation does not fail, although I'm afraid more complexity is required).
I understand that adding this kind of control to all the possible data formats will be definitely complex and require a lot of additional work, so I will not request this for this review. However I suggest you consider some kind of control mechanism for future developments.
