### 2024-09-26 MillionConcepts/pdr#68 @m-stclair
You make a good point. We perform this kind of check on binary tables/arrays (it's fully built into how we read them), but not ASCII. Basically this is because we have found that people often play very fast and loose with data type specifications for columns/fields in ASCII files. However, it might be time to revisit this. This is not specific to the LRO products, so I have opened [a separate issue](https://github.com/MillionConcepts/pdr/issues/73) if you would like to discuss further there.

Leaving this issue open in case Gael gets a response from the data provider.
