### 2024-09-26 MillionConcepts/pdr#68 @Sierra-MC
@gaelccc I've created a full list of all affected ku2s files in the geo holdings. It is possible there are some other wea files that have this issue, but I have only checked the ku2s to forego having to download the entire corpus on my laptop. I'm not sure who you contacted already but perhaps this list would be helpful to them.

[affected_files.csv](https://github.com/user-attachments/files/17155152/affected_files.csv)
