### 2024-09-29 #7256 @dfm
@AndrewAnnex, @gaelccc — I see that you have completed your checklists. Thanks for your quick, thorough, and constructive reviews!! Please let me know if there are any remaining issue that you would like to address before publication of this paper.

@Sierra-MC — Thanks for all your work responding to the comments from the reviewers! I'm sorry that I wasn't able to respond to your questions when you asked them. (I can only work on JOSS on the weekends these days unfortunately.)

> @dfm I don't see a style guide on the website, does JOSS have one?

We don't have a strict style guide for the papers, and I'm happy with the manuscript as written. Thanks!

While we wait to hear confirmation of approval from @AndrewAnnex and @gaelccc, I have a few last steps for you:

2. Increment the version number of the software and report that version number back here. The goal here is to associate a version number with the state of the software at the end of the review.
