+++ CHANGELOG.md
@@ -1,648 +1,0 @@
-## [1.2.2] - 2024-XX-XX
-### Added
-
-#### Features
-- Support for 'desktop' image formats in primary mode: JPEG, TIFF, GIF 
-  (still and animated), PNG, MPO/MPF, WebP (still and animated), JPEG2000/JP2, 
-  and BMP. Behavior is as follows:
-  - `pdr` will attempt to parse all EXIF, TIFF tag, and MPO metadata, including
-  embedded XML packets. Parsed metadata is available in `Data.metadata` and 
-  accessible with `Data.metaget()` and related methods.
-  - Some basic image information is also available in `Data.metadata` (format, 
-  color mode, dimensions, MIME type, etc.) 
-  - Most files will have a single data object named 'IMAGE'.
-  - Animated GIF and WebP images will have one data object per frame, named 
-    'FRAME_0', 'FRAME_1', etc.
-  - MPO/MPF files (this includes many files with .jpg and .jpeg extensions)
-  will have one data object per image. The MPO primary image is named 'IMAGE'. 
-  Subsequent images are assigned names based on their MPType and index within
-  the file.
-  - PDR automatically converts palette-mode images (like many GIFs) to 3D RGB
-  arrays on load. In these cases, mappings between RGB values and original 
-  palette indices are available with `data.metaget('palette')`.
-  - GeoTiffs work like other TIFF files. `pdr` naively parses the GeoTiff 
-  metadata, but does not construct a CRS or otherwise derive projection 
-  information.
-
-#### Dataset Support
-- additional Cassini ISS calibration images
-- MRO HiRISE RDRs and additional DTM images (were notionally supported, now have official support)
-
-#### Other
-- The text of our JOSS paper publication on pdr is now included under 
-  `joss_paper/paper.md`. It will be moved to `docs/paper.md` in next release.
-- Clarifying comments were added to the supported_datasets documentation regarding
-  spice files
-- Community guidelines including a code of conduct and contribution guidelines
-  have been added to the readme and docs folder
-
-### Changed
-- `Data.dump_browse()` and `.show()` no longer apply default contrast enhancement 
-  if it would result in an image with only one value. Explicitly passing the
-  `image_clip` argument will override this, if for some reason you want to.
-- Levenshtein dependency is now optional. Users installing via pip will need to
-  use the "fuzzy" option.
-
-### Fixed
-- `Data.dump_browse()` now works correctly with FITS files opened in primary mode.
-
-### Removed
-
-
-## [1.2.1] - 2024-09-18
-### Added
-
-#### Features
-- Full compatibility with numpy 2.0 for pds4 products 
-(pds3 products already compatible in last release)
-
-### Changed
-- `pds4_tools` is now vendored within `pdr` to apply changes for numpy 2.0 
-compatibility and avoid hot patches
-
-### Removed
-- `pds4_tools` dependency
-- `_patches` module
-
-## [1.2.0] - 2024-09-18
-### Added
-
-#### Features
-- Compatibility with numpy 2.0 for pds3 products and fits pds4 products
- (non-fits PDS4 product numpy 2.0 compatibility awaits changes in
- `pds4_tools` dependency)
-- Support for VAX_REAL data types in qubes
-- `pdr.Data.__contains__` method
-[see issue #57](https://github.com/MillionConcepts/pdr/issues/57)
-- Unit tests are now backed by `pytest` fixtures for improved flexibility and
-compatibility
-- ROW_PREFIX_BYTES are now accounted for when calculating offset positions
-
-#### Dataset Support
-- Galileo NIMS qubes
-- MSL 
[…]
