### 2024-09-30 cd19b52
switch to jupyterlite demo notebook
