+++ README.md
@@ -46,9 +46,9 @@ and the added functionality they support are listed below:
 
 ### Usage
 
-You can check out our example Notebook on Binder for a 
+You can check out our example Notebook on a JupyterLite server for a 
 quick interactive demo of functionality: 
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/millionconcepts/pdr/main)
+[![JupyterLite](docs/jlitebadge.svg)](https://millionconcepts.github.io/jlite-pdr-demo/)
 
 Additional information on usage including examples, output data types, notes 
 and caveats, tests, etc. can now be accessed in our documentation on 
