### 2024-09-30 aff94e0
ignore meta attributes set to None in some PIL versions
