### 2024-09-30 1c6984f
add tests for primary desktop, light linting
Add tests for primary-mode desktop images. Also performed Light linting on some other tests.
