### 2024-09-30 95ec3fd
Add data for primary image tests and change optional dependency name for accuracy.
Added all data for primary-mode desktop image tests. Updated changelog.
Changed name of optional "tiff" dependency category to "desktop_image"
to more accurately reflect current functionality.
