### 2024-09-30 717cb8a
add primary desktop test images to package
