### 2024-09-30 74111c0
update ci and changelog for changed feature name
