### 2024-09-30 aa32846
attempt to package the data files
