### 2024-09-30 55499f0
remove setuptools-scm due to too much being packaged
