### 2024-09-30 61b7a7f
add test data to MANIFEST so it is included in sdist
