### 2024-09-30 f1dec31
include version_history and specify which env files in MANIFEST
