+++ MANIFEST.in
@@ -1,6 +1,7 @@
 # No, it isn't possible to do this in pyproject.toml :-(
-include CHANGELOG.md
-include *environment.yml
+include docs/version_history.md
+include environment.yml
+include minimal_environment.yml
 include .coveragerc
 recursive-include tests *.py
 include tests/data/*
