### 2024-09-30 5975e9e
format pip optionals nicely
