### 2024-09-30 83fa358
assorted dependency cleanup
Removed save_sparklines() and, along with it, the optional matplotlib
dependency. Consolidated optional dependency groups that contained only
pillow. Added useful errors for attempts to use functionality that
requires an optional but not present dependency.
