### 2024-09-30 038a300
special cases for MGS MOLA pedr tables
