### 2024-09-30 1f471b2
fix jlite badge for index.md
