### 2024-09-30 9d0cafa
bump version number; update api doc
