+++ docs/version_history.md
@@ -1,5 +1,23 @@
 # Version History
 
+## [X.X.X] - 202X-XX-XX
+### Added
+
+#### Features
+
+
+#### Dataset Support
+
+
+### Changed
+
+
+### Fixed
+
+
+### Removed
+
+
 ## [1.2.3] - 2024-09-30
 
 ### Changed
