### 2024-09-30 #7256 @Sierra-MC
Okay! Thank you all so much! @AndrewAnnex and @gaelccc thank you especially for the constructive and helpful review experience.

 I have reviewed the manuscript and am happy with it. I've incremented the version number and made a new release. This is version 1.2.3 (such a fun combo). The archived release of this version with exactly matching title and metadata is here on zenodo: https://doi.org/10.5281/zenodo.13864063

Thank you very much @dfm for your time on this.
