### 2024-09-30 e93d870
don't reset the index for mgs mola pedr tables
These products are really one large table, with different ways of reading bytes 501-528 depending on the pointer/format file. By not resetting the index, the user can more easily recombine each pointers' output back into one large table
