### 2024-09-30 #7256 @dfm
@AndrewAnnex, @gaelccc — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

@Sierra-MC — Can you fix the "invalid DOI" indicated by the bot above?
