### 2024-10-01 #7256 @Sierra-MC
@dfm Confirmed this DOI issue has been fixed. Thank you.
