### 2024-10-02 cd1faa2
change strict pointer name check to substring check when deduping pointers
