### 2024-10-02 e4d11c1
Move the unit tests back into the pdr package.
After some internal discussion, we *do* want to ship the unit tests
in the wheel, so put them back into the pdr packge so we can do that.
This has the nice side effect of simplifying the package metadata and
the CI scripts.
