### 2024-10-02 832192f
Need to mention tests/data/* in MANIFEST.in.
This seemed to be unnecessary locally but more stringent testing reveals
it's needed after all.
