### 2024-10-02 4528b7f
Change how MGS MOLA pedr tables are given to the user
Before: each pointer returned a subset of the table, only rows with a FRAME_INDEX flag that matched the pointer name.
Now: the whole table is returned regardless of pointer used. The difference in table outputs is which format file was applied to the engineering data columns. (This is more in line with how the products are described in their labels and documentation.)
