### 2024-10-03 f837b74
Use check_special_block with mgs_mola pedr
parse_pvl's soft check for substrings invalidates the need for earlier
special case checks. change mgs_mola special case to use
check_special_block instead.
