### 2024-10-03 620a9f1
delete the zombie mola_pedr_table_loader() special case
