### 2024-10-03 ce2c537
reindex DESCRIPTION pointers; trivialize PDS_OBJECT pointers
