### 2024-10-04 7248f7c
some edits to browsification docs
