### 2024-10-04 2ef1aa3
useful message for importerror on primary desktop image read
