### 2024-10-05 #7256 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
