### 2024-10-05 #7256 @warrickball
Hi @Sierra-MC! :wave: I have a few small fixes ahead of starting the publication process.

* It's exceptional to use **strong emphasis** in JOSS text, so please change "**data-driven development**" to "*data-driven* development". (Note that the emphasis is only on the development being *data-driven*.)
* The Astropy project's [website advises](https://www.astropy.org/acknowledging.html) (as also indicated in their [CITATION.cff](https://citation-file-format.github.io/) file) the preferred citation being at least AstroPy [Paper III](https://ui.adsabs.harvard.edu/abs/2022ApJ...935..167A), if not also papers [I](http://adsabs.harvard.edu/abs/2013A%26A...558A..33A) and [II](http://adsabs.harvard.edu/abs/2018arXiv180102634T). We'd support you citing all three but you should at least cite Paper III instead of the GitHub repo.
* Though the title of Kaufman et al. (2022) is all caps in the PDF of the proceedings, this appears (from the conference program) to be style, so the article title should be in sentence case (i.e. "How We’ll Know We Can Read All the Data in the PDS:  A Testing Methodology for the Planetary Data Reader!!!").
* Rasterio also has a [recommended citation](https://github.com/rasterio/rasterio/blob/main/CITATION.txt) that is slightly different from what you've used in the paper.

Once these are fixed we'll be ready to publish. Thanks!
