+++ pdr/parselabel/utils.py
@@ -2,8 +2,6 @@ from pathlib import Path
 import re
 from typing import Union, IO
 
-from pdr.utils import head_file
-
 
 KNOWN_LABEL_ENDINGS = (
     re.compile(b"\nEND {0,2}(\r| {8})"),  # common PVL convention
@@ -18,16 +16,47 @@ DEFAULT_PVL_LIMIT = 1000 * 1024
 """heuristic for max label size. we know it's not a real rule."""
 
 
+class InvalidAttachedLabel(ValueError):
+    pass
+
+
+def _scan_to_end_of_label(buf: IO, max_size: int, text: bytes):
+    """Subroutine of trim_label()"""
+    length = 0
+    while length < max_size:
+        # TODO, maybe: this will perform null reads over and over on short
+        #  files with no attached label. This is probably not a big deal, but
+        #  we could add a check.
+        chunk = buf.read(50 * 1024)
+        for ending in KNOWN_LABEL_ENDINGS:
+            if (endmatch := re.search(ending, text[:-15] + chunk)) is not None:
+                return text + chunk[: endmatch.span()[1]]
+        text, length = text + chunk, length + 50 * 1024
+    raise InvalidAttachedLabel("Couldn't find a label ending.")
+
+
 def trim_label(
     fn: Union[IO, Path, str],
     max_size: int = DEFAULT_PVL_LIMIT,
-    raise_for_failure: bool = False,
-) -> bytes:
+    strict_decode: bool = True
+) -> str:
     """Look for a PVL label at the top of a file."""
-    head = head_file(fn, max_size).read()
-    for ending in KNOWN_LABEL_ENDINGS:
-        if (endmatch := re.search(ending, head)) is not None:
-            return head[: endmatch.span()[1]]
-    if raise_for_failure:
-        raise ValueError("couldn't find a label ending")
-    return head
+    target_is_fn = isinstance(fn, (Path, str))
+    try:
+        if target_is_fn is True:
+            fn = open(fn, 'rb')
+        text = fn.read(20)
+        if strict_decode is True:
+            try:
+                text.decode('ascii')
+            except UnicodeDecodeError:
+                raise InvalidAttachedLabel("File head appears to be binary.")
+        text = _scan_to_end_of_label(fn, max_size, text)
+    finally:
+        if target_is_fn is True:
+            fn.close()
+    policy = "strict" if strict_decode is True else "replace"
+    try:
+        return text.decode("utf-8", errors=policy)
+    except UnicodeDecodeError:
+        raise InvalidAttachedLabel("Invalid characters in label.")
