### 2024-10-06 e97f610
add former permissive no-ending behavior back, stop reading when file is over
