+++ joss_paper/paper.bib
@@ -116,11 +116,66 @@
       title = "{The Astropy Project}"
 }
 
-@misc{rasterio,
-      author = {Mapbox},
-      url = {https://github.com/rasterio/rasterio/tree/main},
-      publisher = {Github},
-      journal = {Github repository},
-      year = 2024,
-      title = "{Rasterio}"
+@article{astropy1,
+         author = {{Astropy Collaboration} and {Robitaille}, Thomas P. and {Tollerud}, Erik J. and {Greenfield}, Perry and {Droettboom}, Michael and {Bray}, Erik and {Aldcroft}, Tom and {Davis}, Matt and {Ginsburg}, Adam and {Price-Whelan}, Adrian M. and {Kerzendorf}, Wolfgang E. and {Conley}, Alexander and {Crighton}, Neil and {Barbary}, Kyle and {Muna}, Demitri and {Ferguson}, Henry and {Grollier}, Fr{\'e}d{\'e}ric and {Parikh}, Madhura M. and {Nair}, Prasanth H. and {Unther}, Hans M. and {Deil}, Christoph and {Woillez}, Julien and {Conseil}, Simon and {Kramer}, Roban and {Turner}, James E.~H. and {Singer}, Leo and {Fox}, Ryan and {Weaver}, Benjamin A. and {Zabalza}, Victor and {Edwards}, Zachary I. and {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, Andrew R. and {Crawford}, Steven M. and {Dencheva}, Nadia and {Ely}, Justin and {Jenness}, Tim and {Labrie}, Kathleen and {Lim}, Pey Lian and {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
+         title = "{Astropy: A community Python package for astronomy}",
+         journal = {\aap},
+         keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2013,
+         month = oct,
+         volume = {558},
+         eid = {A33},
+         pages = {A33},
+         doi = {10.1051/0004-6361/201322068},
+         archivePrefix = {arXiv},
+         eprint = {1307.6212},
+         primaryClass = {astro-ph.IM},
+         adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..33A},
+         adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{astropy2,
[…]
