### 2024-10-08 1c01737
fix reference rendering issue
