### 2024-10-08 #7256 @Sierra-MC
Hi @warrickball,

Thank you very much for your feedback. Apologies for my delayed response, I was out of the office. The changes you requested for the strong emphasis, and the references for astropy and rasterio have been made.

For the Kaufman et al., 2022 citation the title is intended to be in all caps. That's an abstract/poster I was first author on (before changing my last name); here is a screenshot of the poster itself as well displaying the title.

![Screenshot from 2024-10-08 12-03-03](https://github.com/user-attachments/assets/6e6808ce-1961-412d-80ce-04ba09a0cacf)
