+++ pyproject.toml
@@ -31,6 +31,7 @@ dependencies = [
     "multidict",
     "numpy",
     "pandas>=2.0.0",
+    "rms-vax"
 ]
 
 [project.optional-dependencies]
