### 2024-10-08 #7256 @warrickball
@Sierra-MC, thanks! And if that's the intended title, no problem. I only saw the sentence case in the program but it's your poster and you get the final say. I'll re-run the pre-publication step but I expect we'll be ready to proceed.
