### 2024-10-08 #7256 @warrickball
Two more things:

* Could you please expand the addresses in the affiliations, including the country? (I apologise that I should've picked this up before.)
* The Astropy citations you've used probably use the journal shortcuts like `\apj` instead of *The Astrophysical Journal*, which JOSS doesn't expand, leading to a blank field for the journal name. I think it's only those three new citations that have this issue but you may wish to double check the others, too.
