### 2024-10-08 6c3aab4
add countries and journal names
