### 2024-10-08 9f174d5
Explicitly identify given names and surnames to fix JOSS metadata
