### 2024-10-09 #7256 @warrickball
Thanks once again to @AndrewAnnex and @gaelccc for reviewing and @dfm for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.
